requirements.txt
alembic==1.5.2
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
